Add methods to get data from archive
Find Bus Open Data Service has 2 links that allows users to
download zip archives containing a dump of all the AVL data, one
in the SIRI VM format and the other in GTFS Realtime.

I've added methods fetch these zip files, extract the data within
the zip and return the data.

For SIRI VM this is a byte string that can be used by lxml and
for GTFS RT a FeedMessage object is returned.

Upgraded black to a more recent version and fixed formatting
issues.