Add Siri Pydantic models
The client currently returns an XML byte string.

In a future update I would like to start returning python objects.
This will require the XML byte string to be parsed.
To do this I'm using a combination of `pydantic` and `lxml`.

I've included tests with various response types.

This change requires the addition of `lxml` as a dependency.